cors origin policy update v3

+++ api/index.js
@@ -70,7 +70,7 @@ async function uploadToS3(path, originalFilename,mimetype){
 }
 
 app.get('/api/test',(req,res)=>{
-    res.json({message:"hello world"})
+    res.json({message:"hello world"}).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
 })
 
 
@@ -84,12 +84,12 @@ app.post('/api/register',async(req,res)=>{
             email,
             password:bcrypt.hashSync(password, bcryptSalt)
         });
-        res.json(userDoc)
+        res.json(userDoc).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
 
 
     } catch (error) {
         
-        res.status(422).json(error)
+        res.status(422).json(error).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
     }
     
 })
@@ -100,7 +100,7 @@ app.post('/api/login',async(req,res)=>{
     const userDoc =await User.findOne({email:email})
     
     if(!userDoc){
-        res.status(422).json({message:"email is incorrect"})
+        res.status(422).json({message:"email is incorrect"}).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
     }
     if(userDoc){
         
@@ -113,11 +113,11 @@ app.post('/api/login',async(req,res)=>{
                 process.env.JWTSECRET, 
                 {}, (err,token)=>{
                 if (err) throw err
-                res.cookie('token',token).json(userDoc)
+                res.cookie('token',token).json(userDoc).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
             })
             
         }else{
-            res.status(422).json({message:"password is incorrect"})
+            res.status(422).json({message:"password is incorrect"}).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
         }
     }
 })
@@ -131,14 +131,14 @@ app.get('/api/profile', async(req,res)=>{
         res.json(userData)
         
     }else{
-        res.status(401).json({message:"unauthorized"})
+        res.status(401).json({message:"unauthorized"}).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
     }  
 })
 
 
 app.post('/api/logout', (req,res)=>{
 
-    res.clearCookie('token').json(true)
+    res.clearCookie('token').json(true).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
 })
 
 
@@ -151,7 +151,7 @@ app.post('/api/upload-with-link',async (req,res)=>{
         dest:'/tmp/'+ name
     });
     const url = await uploadToS3('/tmp/'+ name, name, mime.lookup('/tmp/'+ name))
-    res.json(url)
+    res.json(url).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
 })
 
 const pictureMiddleware= multer({dest:'tmp'})
@@ -166,7 +166,7 @@ app.post('/api/upload',pictureMiddleware.array('pictures',100) ,async(req,res)=>
         
 
     }
-    res.json(uploads)
+    res.json(uploads).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
 })
 
 app.post("/api/place", (req,res)=>{
@@ -188,7 +188,7 @@ app.post("/api/place", (req,res)=>{
             description,perks,extraInfo,price,
             checkIn,checkOut,maxGuests
         })
-        res.json(placeDocs)
+        res.json(placeDocs).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
     })
     
 })
@@ -200,7 +200,7 @@ app.get('/api/user-places', (req,res)=>{
     jwt.verify(token,process.env.JWTSECRET,{},async(err,decoded)=>{
         const {id}= decoded;
         const places = await Place.find({owner:id})
-        res.json(places)
+        res.json(places).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
     })
 })
 
@@ -209,7 +209,7 @@ app.get('/api/user-places', (req,res)=>{
 app.get('/api/places/:id', async (req,res)=>{
     mongoose.connect(uri, {useNewUrlParser: true, useUnifiedTopology: true} );
     const {id}= req.params
-    res.json(await Place.findById(id))
+    res.json(await Place.findById(id)).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
 })
 
 
@@ -231,7 +231,7 @@ app.put('/api/place',(req,res)=>{
             checkIn,checkOut,maxGuests
         })
    
-        res.json(placeDoc)
+        res.json(placeDoc).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
         
     })
 
@@ -242,13 +242,13 @@ app.delete('/api/place/:id',async(req,res)=>{
     const {id}= req.params
     console.log(id)
     const userData=await getUserDataFromCookies(req)
-    res.json(await Place.deleteOne({_id:id, owner:userData.id}))
+    res.json(await Place.deleteOne({_id:id, owner:userData.id})).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
 })
 
 
 app.get('/api/places',async(req,res)=>{
     mongoose.connect(uri, {useNewUrlParser: true, useUnifiedTopology: true} );
-    res.json(await Place.find())
+    res.json(await Place.find()).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
 })
 
 
@@ -263,7 +263,7 @@ app.get('/api/room/:id', async (req,res)=>{
     }
 
     const place= await Place.findById(id)
-    res.json([place,dates])
+    res.json([place,dates]).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
 });
 
 
@@ -275,7 +275,7 @@ app.post('/api/booking',  async(req,res)=>{
 
     Booking.create({place,checkIn,checkOut,numberOfGuests,name,phone,price,user:userData.id
     }).then((doc)=>{
-        res.json(doc)
+        res.json(doc).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
     }).catch((err=>{
         console.error(err)
     }))
@@ -288,7 +288,7 @@ app.get('/api/bookings',  async (req,res)=>{
     
     const response= await Booking.find({user:userData.id}).populate('place')
     
-    res.json(response)
+    res.json(response).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
 
 })
 
@@ -298,7 +298,7 @@ app.delete('/api/booking/:id',async(req,res)=>{
     const {id}= req.params
     console.log(id)
     const userData=await getUserDataFromCookies(req)
-    res.json(await Booking.deleteOne({_id:id, user:userData.id}))
+    res.json(await Booking.deleteOne({_id:id, user:userData.id})).setHeader("Access-Control-Allow-Origin", "https://back-booking-app.onrender.com")
 })
